rename settings -> pumpSettings
