+++ js/validation/common.js
@@ -24,6 +24,6 @@ module.exports = schema(
     joinKey: schema().ifExists().isId(),
     normalTime: schema().isISODateTime(),
     time: schema().isISODateTime(),
-    type: schema().string().in(['basal', 'bolus', 'cbg', 'message', 'smbg', 'settings', 'wizard'])
+    type: schema().string().in(['basal', 'bolus', 'cbg', 'message', 'smbg', 'pumpSettings', 'wizard'])
   }
 );
